+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.4.8
+- Virtual Pet Hud Statistic tooltip feature by dDeepLb
+- Add Echo's Clothing Extenions beds to energy regain list
+- Virtual Pet HUD override option; Will display even with BCX rule "Force-hide UI elements"
+- Miscellaneous bug fixes & Stability improvements
+- Localization improvements
+- Stomp activity while wearing pony boots
+- Add Chinese localization thanks to Da'Inihlus 
+- Bell jingle on movement (emote, map, or pose); Thanks to Alexa (Corvidae) for the audio
+
 # 0.4.7
 - Support clicker for activities (including Echo's custom activities)
 - RU localization by rubich9141